WIP - Simple HTML Rendering, but problem with framing
I'm running into an issue when there's a Section heading followed
immediately by a Subsection heading. My strategy of framing and then
encoding separately might not work as well as I think.

Need to figure this out. I could potentially add another thing
in the pipeline, if I wanted to.