Cleaned up samples, reworked framing, started serde
Instead of having multiple pipeline entries for framing,
I now have everything happening in one go. The code is
simple enough that I'm OK with it.

I've also decided that I'm going to try to make all of
the classes serializeable, and I've started with frames.
I think this is going to help when I swap parts of this
out for Go in the future. I can really just stick the
thing together with unix pipes.

I'm also getting closer to a stable spec wrt links. Not
there yet.