Reworked block decoders, remove config.py, added NoteElement
Block decoders now all live in blocks instead of being instances of
BlockDecoder in config.py. I think this will add flexibility in the
future as I finish code handling, etc.

ParagraphDecoder now considers a paragraph with lines == ['.'] as
a noop - It's just a way to signify a break. See previous commit.